requirements.txt
python-telegram-bot
flask
flask-login
requests
sqlalchemy
psycopg2
numpy
matplotlib
scipy
opencv-python
pytest
pytest-cov
dropbox
click
click-repl
faker
keras
flask-wtf
tensorflow
keras
scikit-image
python-google-places
plotly

